Clean up content manager configuration code and init tests
