+++ packages/strapi-plugin-content-manager/services/utils/configuration/__tests__/attributes.test.js
@@ -1,0 +1,15 @@
+const {
+  isSortable,
+  // isEditable,
+  // hasEditableAttribute,
+  // hasListableAttribute,
+  // hasRelationAttribute,
+} = require('../attributes');
+
+describe('attributesUtils', () => {
+  describe('isSortable', () => {
+    test('The id attribute is always sortable', () => {
+      expect(isSortable({}, 'id')).toBe(true);
+    });
+  });
+});
